requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
dateparser==1.1.4
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.4
google-cloud-bigquery-logging==1.0.7
google-cloud-core==2.3.2
google-cloud-logging==3.3.1
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
oauth2client==4.1.3
pkg_resources==0.0.0
proto-plus==1.22.1
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
rsa==4.9
six==1.16.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
